requirements.txt
aiogram==3.20.0.post0
pyotp==2.9.0
requests==2.31.0
python-dotenv==1.0.0
selenium==4.9.0